updated type for id columns from integer to bigint

+++ Backend-Bot-master/app/models/car_photo.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String, TIMESTAMP, func, Text
+from sqlalchemy import BigInteger, String, TIMESTAMP, func, Text
 from sqlalchemy.orm import Mapped, mapped_column
 from app.db import Base
 from sqlalchemy import ForeignKey
@@ -7,8 +7,8 @@ from sqlalchemy.orm import relationship
 class CarPhoto(Base):
     __tablename__ = 'car_photos'
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    car_id: Mapped[int] = mapped_column(Integer, ForeignKey('cars.id'), nullable=False)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    car_id: Mapped[int] = mapped_column(BigInteger, ForeignKey('cars.id'), nullable=False)
     type: Mapped[str] = mapped_column(String(50), nullable=True)
     description: Mapped[str] = mapped_column(String(255), nullable=True)
     photo_url: Mapped[str | None] = mapped_column(String, nullable=True)
