+++ Backend-Bot-master/app/models/role.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String, TIMESTAMP, func, Text
+from sqlalchemy import BigInteger, String, TIMESTAMP, func, Text
 from sqlalchemy.orm import Mapped, mapped_column
 from app.db import Base
 
@@ -6,7 +6,7 @@ from app.db import Base
 class Role(Base):
     __tablename__ = 'roles'
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     code: Mapped[str] = mapped_column(String(50), nullable=False, unique=True)
     name: Mapped[str] = mapped_column(String(100), nullable=True)
     description: Mapped[str | None] = mapped_column(Text, nullable=True)
